Update all plugins for Dalamud API 15 (Patch 7.5)
- Bump DalamudApiLevel 14 -> 15 across all 4 csprojs + pluginmaster.json
- Bump assembly versions (CalloutPlugin 1.0.0.4; Strat/PullReady/Currency 1.0.0.1)
- Fix TerritoryChanged delegate: ushort -> uint signature change
- Inject IObjectTable so LocalPlayer access works (moved from IClientState in API 15)
- Fix CurrencyBoard RepoUrl pointing at non-existent /Thunderz96/CurrencyBoard
- Add ThunderzCombo/ and .claude/ to .gitignore

+++ PullReady/PullReady/Plugin.cs
@@ -144,7 +144,8 @@ public sealed class Plugin : IDalamudPlugin
     /// content) vs. just walking between open-world areas.
     /// BoundByDuty = inside a duty instance (raid, trial, dungeon, etc.)
     /// </summary>
-    private void OnTerritoryChanged(ushort territoryId)
+    // 7.5 / API 15: TerritoryChanged delegate is now Action<uint> (was ushort).
+    private void OnTerritoryChanged(uint territoryId)
     {
         if (!Configuration.AutoOpenOnZoneIn)
             return;
